Rename `extract-new.ts` to `extract.ts`
