+++ lib/modules/manager/bazel/extract.ts
