+++ lib/modules/manager/bazel/index.ts
@@ -2,7 +2,7 @@ import { DockerDatasource } from '../../datasource/docker';
 import { GithubReleasesDatasource } from '../../datasource/github-releases';
 import { GithubTagsDatasource } from '../../datasource/github-tags';
 import { GoDatasource } from '../../datasource/go';
-import { extractPackageFile } from './extract-new';
+import { extractPackageFile } from './extract';
 import { updateDependency } from './update';
 
 export { extractPackageFile, updateDependency };
